WebSocket enhancements
1. The WebSocket session expires after 24h, when this happens Avanza
will send out a /meta/disconnect after which a new handshake must
be performed before any new operation can take place on the socket.
This code will initiate a new handshake upon reception of
/meta/disconnect or advice.reconnect='handshake'. Note that after
session expiry all subscriptions are lost.

2. In the event the TCP socket drops the 'ws' library will send
an 'close' event after which a new object must be instantiated
to create a new TCP connection. Once this connection is up a
/meta/connect will be sent out and subscriptions will continue
to function.